+++ src/graphics/basicEffect.ts
@@ -1,7 +1,7 @@
-import shaders from "./shaders.js";
+import shaders from "./geometry/polygonShaders.js";
 import Effect from "./effect.js";
 // eslint-disable-next-line no-unused-vars
-import Graphics from "./graphics.js";
+import Graphics from "./graphicsManager.js";
 
 export default class BasicEffect extends Effect {
 	constructor(graphics: Graphics) {
