+++ src/graphics/buffer.ts
@@ -1,6 +1,6 @@
 import * as WebGL from "./webGL.js";
 // eslint-disable-next-line no-unused-vars
-import Graphics from "./graphics.js";
+import Graphics from "./graphicsManager.js";
 import BufferType from "./bufferType.js";
 
 export default class Buffer {
